Made schedule appointment it's own component so it could be reused
